% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CalcFst}
\alias{CalcFst}
\title{Calculate Fst using the model}
\usage{
CalcFst(results)
}
\arguments{
\item{results}{- the output of \code{\link{sample.bayesFst}}}
}
\value{
a matrix with \code{nout} rows and \code{nloci * npop} columns
}
\description{
Calculate Fst using the model
}
