% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Intialise a new Bayes Fst Object}
\usage{
init(data)
}
\arguments{
\item{data}{a dataset of class \code{bayesFstData} loaded by \code{readData}.}
}
\value{
an object of class \code{bayesFst}
}
\description{
Intialise a new Bayes Fst Object
}
\examples{
bd = readData(system.file("extdata", "data_BB04.json", package = "rbayesfst"))
bf = init(bd)
}

