% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bayesFst.R
\name{print.bayesFst}
\alias{print.bayesFst}
\title{Print method for \code{bayesFst} objects}
\usage{
\method{print}{bayesFst}(x, ...)
}
\arguments{
\item{x}{An object of class \code{bayesFstData}.}

\item{\ldots}{Not used.}
}
\description{
Provides a simple summary of the input data used with this package.
}
\seealso{
readData
}
