% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printData.R
\name{printData}
\alias{printData}
\title{Print the data bound to an Bayes Fst Object}
\usage{
printData(b)
}
\arguments{
\item{b}{an object of class \code{Rcpp_BayesFst}}
}
\description{
Print the data bound to an Bayes Fst Object
}

