% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readData}
\alias{readData}
\title{readData}
\usage{
readData(fileName)
}
\arguments{
\item{fileName}{}
}
\value{
a dataset of class \code{bayesFst}. This is essentially a list with the following members: \tabular{ll}{
\code{dbCounts} \tab A list of matrices providing allele counts at each locus for each population. The items in the list are loci, the rows are populations, and the columns alleles. \cr
\code{nLoci} \tab the number of loci in the dataset. \cr
\code{Loci} \tab the locus names if available. If the locus names are not available they will be labelled Locus 1, Locus 2, \ldots \cr
\code{nPops} \tab the number of populations in the dataset. \cr
\code{Pops} \tab the population names if available. If the population names are not available they will be labelled Pop 1, Pop 2, \ldots \cr
\code{numAlleles} \tab the number of possible alleles at each locus \cr
\code{locusPopSums} \tab a matrix containing the number of alleles observed at the ith locus for the jth population. \cr
\code{gammaSwitch} \tab either \code{TRUE} or \code{FALSE} depending on whether locus and popuation effects interact or not \cr
}
}
\description{
A function to read the data needed for the algorithm.
}
\details{
A function that can read the data needed for the algorithm. The original program requires data to be in the following format 
Input format

The input file should be a space or tab-delimited ASCII (i.e. text)
file. First, there are three header lines, containing:

\enumerate{
\item interaction term (gamma) switch: 0 - for no interaction term 1 - to include interaction term
\item the number of subpopulations (npop)
\item the number of loci (nloc)
}

Then for locus i (i=1,..,nloc) there are npop+1 input lines, the first
containing the number of alleles at that locus (numall[i]) and the
remaining lines each contain the allele counts.

For example, for 3 subpops and 2 loci, with 2 and 4 alleles, where we
do not want to include the interaction gamma in the analysis, the input
should look like:
  
\code{0} \cr
\code{3} \cr
\code{2} \cr


\code{2} \cr
\code{45 5} \cr
\code{23 27} \cr
\code{10 40} \cr
 \cr
\code{4} \cr
\code{10 10 10 20} \cr
\code{5 5 25 15} \cr
\code{17 19 0 14} \cr
Spaces and blank lines can be included as desired. Unlike FSTMET.C, in this
program it is assumed that there is no data directly sampled from the "total"
population.
}

