% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample.bayesFst}
\alias{sample.bayesFst}
\title{Sample from the posterior distribution of Fst}
\usage{
\method{sample}{bayesFst}(x, size = NULL, seed = floor(runif(1, 0, 2^32) + 1), ...)
}
\arguments{
\item{x}{an object class \code{bayesFst}}

\item{size}{the desired numbers of samples to be taken from the posterior distribution of the parameters}

\item{seed}{a positive (32 bit) integer to be used as a random number seed in a Mersenne-Twister random number generator. This package uses the C++ `random` library.}

\item{\ldots}{not used}
}
\value{
a list
}
\description{
Sample from the posterior distribution of Fst
}
\note{
If you want to change some of the parameters for the MCMC algorithm, other than the number of outputs, then use the specific methods to do this. They cannot be changed here.
}
\seealso{
\code{\link{setInteraction}}, \code{\link{setPriors}}, \code{\link{setRunParams}}
}
