% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample.default}
\alias{sample.default}
\title{Default call to base R \code{\link[base]{sample}} function}
\usage{
\method{sample}{default}(x, size, replace = FALSE, prob = NULL, ...)
}
\arguments{
\item{x}{either a vector of one or more elements from which to choose,
or a positive integer.}

\item{size}{a non-negative integer giving the number of items to choose.}

\item{replace}{should sampling be with replacement?}

\item{prob}{a vector of probability weights for obtaining the elements of 
the vector being sampled.}

\item{n}{a positive number, the number of items to choose from.}

\item{\ldots}{any other arguments. Not used.}
}
\description{
Default call to base R \code{\link[base]{sample}} function
}
\seealso{
\code{\link[base]{sample}}
}
