% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample}
\alias{sample}
\title{sample generic}
\usage{
sample(x, ...)
}
\arguments{
\item{x}{an object.}

\item{\dots}{Any additional arguments to be passed to \code{sample}.}
}
\description{
sample generic
}

