% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setInteraction.R
\name{setInteraction}
\alias{setInteraction}
\title{Change the model to use (or not use) the interaction term}
\usage{
setInteraction(x, bInteraction = TRUE)
}
\arguments{
\item{x}{an object of class \code{bayesFst}}

\item{bInteraction}{if \code{TRUE} then the interaction (gamma) terms are used in the model, otherwise they are not.}
}
\note{
This function changes a value in a C++ object. As a consequence it DOES NOT need to be reassigned.
}
\examples{
bd = readData(system.file("extdata", "data_BB04.json", package = "rbayesfst"))
bf = init(bd)
print(bf)
## turn the interaction term on
setInteraction(bf)
## note how the model has changed even though there is no reassignment
print(bf)
}

