% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPriors.R
\name{setPriors}
\alias{setPriors}
\title{Set the values of the parameters for the priors}
\usage{
setPriors(
  x,
  alphaMu = 0,
  alphaSigma = 1,
  betaMu = -2,
  betaSigma = 1.8,
  gammaMu = 0,
  gammaSigma = 0.5,
  uSigma = 0.5,
  pSigma = 1000,
  cor = 0
)
}
\arguments{
\item{x}{an object of type bayesFst}

\item{alphaMu}{(expected) mean of Normal distribution for prior of the locus effect alpha[i]}

\item{alphaSigma}{(expected) mean of Normal distribution for prior of locus effect alpha[i]}

\item{betaMu}{mean of Normal distribution for prior of the population effect beta[j]}

\item{betaSigma}{sd of Normal distribution for prior of the population effect  beta[j]}

\item{gammaMu}{mean of Normal distribution for prior of the locus x population interaction effect gamma[i,j]}

\item{gammaSigma}{sd of Normal distribution for prior the locus x population interaction effect of gamma[i,j]}

\item{uSigma}{scale param for Normal updates}

\item{pSigma}{scale param for Dirichlet updates of the allele frequencies p[i,j]}

\item{cor}{The correlation parameter (fixed) between adjacent loci}
}
\description{
Set the values of the parameters for the priors
}
\note{
This function directly affects a C++ object and therefore DOES NOT return anything.
}
\examples{
bd = readData(system.file("extdata", "data_BB04.json", package = "rbayesfst"))
bf = init(bd)
print(bf)
## change the alpha mean only
setPriors(bf, alphaMu = 1)
## note how the mean has changed even though there is no reassignment
print(bf)
}
\seealso{
getPriors
}
