% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setRunParams.R
\name{setRunParams}
\alias{setRunParams}
\title{Set the parameters for the MCMC run}
\usage{
setRunParams(
  x,
  numOut = 2001,
  keepPpn = 0.2,
  discardPpn = 0.05,
  acceptPpn = 0.02,
  print = FALSE
)
}
\arguments{
\item{x}{an object of class bayesFst}

\item{numOut}{the desired numbers of samples to be taken from the posterior}

\item{keepPpn}{\eqn{\pi_k} a thinning proportion used to calculate the interval between successive outputs \eqn{k = \pi_k(n_P+n_L)} where \eqn{n_P} is the number of populations and \eqn{n_L} is the number of loci. This number is in turn used to calculate the number of interations after burn-in \eqn{n_{iter}=k\times n_out}}

\item{discardPpn}{\eqn{\pi_{discard}} the proportion of iterations to be used as burn-in used to calculate \eqn{d = n_{iter}\pi_d}. The total number of iterations is \eqn{d+n_{iter}}}

\item{acceptPpn}{\eqn{\pi_a} used to calculate the gap between output of acceptance rates \eqn{a = \pi_a(d+n_iter)}}

\item{print}{controls the output of something but I forget what at this point}
}
\description{
Set the parameters for the MCMC run
}
\note{
This function sets values in a C++ object. As a consequence it DOES NOT need to be re-assigned.
}
\examples{
bd = readData(system.file("extdata", "data_BB04.json", package = "rbayesfst"))
bf = init(bd)
## set things up for a short test run
setRunParams(bf, numOut = 20)
}
