% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bayesFstData.R
\name{summary.bayesFstData}
\alias{summary.bayesFstData}
\title{Summary information for \code{bayesFstData} objects}
\usage{
\method{summary}{bayesFstData}(object, ...)
}
\arguments{
\item{object}{An object of class \code{bayesFstData}.}

\item{\ldots}{Not used.}
}
\description{
Provides a simple summary of the input data used with this package.
}
\seealso{
readData
}
