% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample.default}
\alias{sample.default}
\title{Default call to base R \code{\link[base]{sample}} function}
\usage{
\method{sample}{default}(x, size, replace = FALSE, prob = NULL)
}
\description{
Default call to base R \code{\link[base]{sample}} function
}
\seealso{
\code{\link[base]{sample}}
}

