% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bayesFstData.R
\name{summary.bayesFstData}
\alias{summary.bayesFstData}
\title{An S3 summary method for bayesFstData objects}
\usage{
\method{summary}{bayesFstData}(x, ...)
}
\arguments{
\item{x}{An object of class \code{bayesFstData}.}

\item{\ldots}{Not used.}
}
\description{
Provides a simple summary of the input data used with this package.
}

